Выделяет шаблоны для отрисовки в отдельные файлы
