+++ src/templates/films-container-template.js
@@ -1,0 +1,1 @@
+export const createFilmsContainerTemplate = () => '<div class="films-list__container"></div>';
