+++ src/templates/films-list-template.js
@@ -1,0 +1,5 @@
+export const createFilmsListTemplate = () => (
+  `<section class="films-list">
+    <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
+  </section>`
+);
