+++ src/templates/films-section-template.js
@@ -1,0 +1,1 @@
+export const createFilmsSectionTemplate = () => '<section class="films"></section>';
