+++ src/templates/movies-amount-template.js
@@ -1,0 +1,4 @@
+export const createMoviesAmountTemplate = (moviesAmount) =>
+  `<section class="footer__statistics">
+    <p>${moviesAmount} movies inside</p>
+  </section>`;
