+++ src/templates/no-movies-template.js
@@ -1,0 +1,7 @@
+export const createNoMoviesTemplate = () => (
+  `<section class="films">
+    <section class="films-list">
+      <h2 class="films-list__title">There are no movies in our database</h2>
+    </section>
+  </section>`
+);
