+++ src/templates/show-more-button-template.js
@@ -1,0 +1,1 @@
+export const createShowMoreButtonTemplate = () => ('<button class="films-list__show-more">Show more</button>');
