+++ src/view/films-container-view.js
@@ -1,6 +1,5 @@
 import { createElement } from '../render.js';
-
-const createFilmsContainerTemplate = () => '<div class="films-list__container"></div>';
+import { createFilmsContainerTemplate } from '../templates/films-container-template.js';
 
 export default class FilmsContainerView {
   #element = null;
