+++ src/view/films-list-view.js
@@ -1,10 +1,5 @@
 import { createElement } from '../render.js';
-
-const createFilmsListTemplate = () => (
-  `<section class="films-list">
-    <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
-  </section>`
-);
+import { createFilmsListTemplate } from '../templates/films-list-template.js';
 
 export default class FilmsListView {
   #element = null;
