+++ src/view/show-more-button-view.js
@@ -1,6 +1,5 @@
 import { createElement } from '../render.js';
-
-const createShowMoreButtonTemplate = () => ('<button class="films-list__show-more">Show more</button>');
+import { createShowMoreButtonTemplate } from '../templates/show-more-button-template.js';
 
 export default class ShowMoreButtonView {
   #element = null;
