+++ src/view/sort-view.js
@@ -1,11 +1,5 @@
 import { createElement } from '../render.js';
-
-const createSortTemplate = () =>
-  `<ul class="sort">
-    <li><a href="#" class="sort__button sort__button--active">Sort by default</a></li>
-    <li><a href="#" class="sort__button">Sort by date</a></li>
-    <li><a href="#" class="sort__button">Sort by rating</a></li>
-  </ul>`;
+import { createSortTemplate } from '../templates/sort-template.js';
 
 export default class SortView {
   #element = null;
