+++ src/view/user-rank-view.js
@@ -1,10 +1,5 @@
 import { createElement } from '../render.js';
-
-const createUserRankTemplate = () =>
-  `<section class="header__profile profile">
-    <p class="profile__rating">Movie Buff</p>
-    <img class="profile__avatar" src="images/bitmap@2x.png" alt="Avatar" width="35" height="35">
-  </section>`;
+import { createUserRankTemplate } from '../templates/user-rank-template.js';
 
 export default class UserRankView {
   #element = null;
